% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrenchSource.R
\name{.getReference}
\alias{.getReference}
\title{This function generates the reference.}
\usage{
.getReference(mat, ref.est = "sw.means", ...)
}
\arguments{
\item{mat}{count matrix; rows are features and columns are samples}

\item{ref.est}{reference estimate method}

\item{...}{other parameters}
}
\value{
the reference to be used for normalization
}
\description{
This function generates the reference.
}
