% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrenchSource.R
\name{.getWeightedMedian}
\alias{.getWeightedMedian}
\title{Get weighted median for matrix}
\usage{
.getWeightedMedian(mat, w = rep(1, nrow(mat)))
}
\arguments{
\item{mat}{input matrix}

\item{w}{weights}
}
\value{
column-wise weighted means.
}
\description{
Get weighted median for matrix
}
