% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrenchSource.R
\name{estimSummary}
\alias{estimSummary}
\title{Obtain robust means. .}
\usage{
estimSummary(res, estim.type = "s2.w.mean", ...)
}
\arguments{
\item{res}{result structure of \code{wrench}}

\item{estim.type}{estimator type}

\item{...}{other parameters}
}
\description{
Obtain robust means. .
}
