% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrenchSource.R
\name{getCondLogWeights}
\alias{getCondLogWeights}
\title{Log Postive-conditional weight computations for wrench estimators.}
\usage{
getCondLogWeights(res)
}
\arguments{
\item{res}{result structure of \code{wrench}}
}
\description{
Log Postive-conditional weight computations for wrench estimators.
}
