% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrenchSource.R
\name{getCondWeights}
\alias{getCondWeights}
\title{Postive-conditional weight computations for wrench estimators.}
\usage{
getCondWeights(res)
}
\arguments{
\item{res}{result structure of \code{wrench}}
}
\description{
Postive-conditional weight computations for wrench estimators.
}

