% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrenchSource.R
\name{getMargWeights}
\alias{getMargWeights}
\title{Marginal weight computations for wrench estimators.}
\usage{
getMargWeights(res, z.adj, ...)
}
\arguments{
\item{res}{result structure of \code{wrench}}

\item{z.adj}{TRUE if the result structure was generated with \code{wrench} with \code{z.adj} set to TRUE.}

\item{...}{other parameters}
}
\description{
Marginal weight computations for wrench estimators.
}
