% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrenchSource.R
\name{gets2}
\alias{gets2}
\title{Obtain variances of logged counts.}
\usage{
gets2(mat, design = model.matrix(mat[1, ] ~ 1), plot = F, ebs2 = T,
  smoothed = F, ...)
}
\arguments{
\item{mat}{count matrix; rows are features and columns are samples.}

\item{design}{model matrix for the count matrix}

\item{plot}{if the mean-variance trend function (the same as that of voom) needs to be plot.}

\item{ebs2}{if regularization of variances needs to be performed.}

\item{smoothed}{TRUE if all the variance estimates must be based on the mean-variance trend function.}
}
\value{
a vector with variance estimates for logged feature-wise counts.
}
\description{
Obtain variances of logged counts.
}

