% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrenchSource.R
\name{getWeightedMean}
\alias{getWeightedMean}
\title{Get weighted means for matrix}
\usage{
getWeightedMean(mat, w = rep(1, nrow(mat)))
}
\arguments{
\item{mat}{input matrix}

\item{w}{weights}
}
\value{
column-wise weighted means.
}
\description{
Get weighted means for matrix
}
